Centralized tests and created builders for the tests.

+++ Pdc.Application/DTOS/ProgramOfStudy/CreateProgramOfStudyDTO.cs
@@ -8,7 +8,7 @@ public class CreateProgramOfStudyDTO
     public required string Code { get; set; } //420.B0
     public required string Name { get; set; } //Techniques de l'informatique
     public required SanctionType Sanction { get; set; } //DEC, PRE-U
-    public double MonthsDuration { get; set; } // 36 mois
+    public int MonthsDuration { get; set; } // 36 mois
     public int SpecificDurationHours { get; set; } // 2010
     public int TotalDurationHours { get; set; } // 5730
     public DateOnly PublishedOn { get; set; }
